// frontend/src/api/generated/api.schemas.ts
export interface Hotel {
  id: number;
  name: string;
  description: string;
  city: string;
  address: string;
  /** @nullable */
  latitude?: number | null;
  /** @nullable */
  longitude?: number | null;
  rating: number;
  stars: number;
  amenities: string[];
  images: string[];
  reviewCount: number;
  /** @nullable */
  minPrice?: number | null;
  createdAt: string;
}